pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bug-investigator"
version = "0.1.0"
description = "LangGraph-based multi-agent automated bug investigation system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=0.2.60",
  "openai>=1.40.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
  "streamlit>=1.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
bug-investigator = "bug_investigator.main:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["bug_investigator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
